Refactor: 컴퓨터 랜덤 숫자 생성 클래스 분리
- RandomGenerator.js
